[pwrmgr] Fix clock stopped asserts
The sequence checking a clock does not advance whilst it is disabled is
automatically false when the clock is only disabled for a single cycle
of the slow clock. This leads to failures where the clock control is
working correctly (namely where at one positive edge of the slow clock a
clock is disabled then at the next positive edge of the slow clock it is
enabled again).

The fix allows an immediate rise of the clock enable after it
fell. If an immediate rise is not seen the clock itself must not have
advanced.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>