[otbn,rtl] Simplify logic driving state_reset
Previously a state reset was performed when the core started or we were
entering RMA. Both of these occur from the core halted state. With this
change we just continuously assert state reset whilst in that state.
This eases the timing on state_reset as it's now just a simple function
of a state lop.